+++ R/dat_checks.R
@@ -6,23 +6,45 @@ dat_check <- function(dat, verbose=TRUE,locuszoomplot=FALSE){
       warning(paste0("Dataset [",i, "] is empty with 0 rows. Please remove it from the input list and re-run!"))
     }
   }
-  dat <- dat_column_check_and_set(dat, verbose=verbose,locuszoomplot=locuszoomplot) %>% dat_chr_check()
+  dat <- dat_column_check_and_set(dat, verbose=verbose,locuszoomplot=locuszoomplot) %>% rm_chr_prefix_and_sort()
   return(dat)
 }
 
-dat_chr_check <- function(dat){
+rm_chr_prefix_and_sort <- function(dat){
   for(i in seq_along(dat)){
     df <- as.data.frame(dat[[i]])
-    #remove chr from CHROM if its there, and set chrX to 23
     df <- df %>% dplyr::mutate(CHROM=gsub('chr','',CHROM))
-    df[df$CHROM=='X', 'CHROM'] <- "23"
-    df$CHROM <-  as.integer(df$CHROM)
-    df <- df %>% dplyr::filter(CHROM<24)
     df$POS <- as.integer(df$POS)
     df <- df %>% dplyr::arrange(CHROM,POS)
     dat[[i]] <- df
   }
   return(dat)
+}
+
+
+convert_chrs_to_numeric <- function(dat, get_chr_lengths_from_data){
+    chrs <- NULL
+    if(get_chr_lengths_from_data)
+      chrs <- get_chrs_from_data(dat)
+    else{
+      chrs <- chr_lengths$V1[2:24] 
+      chrs <- gsub("chr", "", chrs)
+    }
+    numeric_chrs <- chrs[grepl('^-?[0-9.]+$', chrs)] %>% as.numeric() %>% sort()
+    non_numeric_chrs <- chrs[!grepl('^-?[0-9.]+$', chrs)]
+    chr_order <- append(c(numeric_chrs), c(non_numeric_chrs))
+    chr_number=c(1:length(chr_order))
+    chr_map <- stats::setNames(as.list(chr_number), chr_order)
+   for(i in seq_along(dat)){
+     df <- dat[[i]]
+     for(chr in chr_order){
+      df[df$CHROM==chr, 'CHROM'] <- as.numeric(unlist(unname(chr_map[chr])))
+    }
+    df$CRHOM <- as.integer(df$CHROM)
+    df <- df %>% dplyr::arrange(CHROM) 
+    dat[[i]] <- df
+   }
+  return(list("dat"=dat, "chr_map"=chr_map))
 }
 
 convert_region_size <- function(region_size){